TryBindKey: don't write to bindings.json unless overwrite is true
Change the (documented) behavior of TryBindKey(): unless the overwrite
argument is true, never write to bindings.json, even if the new binding
has been successfully applied.

TryBindKey() with overwrite=false is mainly used by plugins for adding
their own default keybindings. It is not really necessary for them to
write those default bindings to bindings.json: even if the binding is
not saved in bindings.json, it will be still applied by the plugin
calling TryBindKey() every time micro starts.

This fixes undesirable behavior: unless there is already a binding for
the given key in bindings.json, a plugin forcibly rewrites
bindings.json, in particular removing any user's comments from it.
With this change, bindings.json will be only rewritten by micro in the
case when the user runs `bind` or `unbind` command, not automatically by
plugins (unless there is a plugin calling TryBindKey() with
overwrite=true, which is unlikely).

+++ internal/action/bindings.go
@@ -256,8 +256,11 @@ func eventsEqual(e1 Event, e2 Event) bool {
 	return e1 == e2
 }
 
-// TryBindKey tries to bind a key by writing to config.ConfigDir/bindings.json
-// Returns true if the keybinding already existed and a possible error
+// TryBindKey tries to bind a key to an action. If there is already a binding
+// for this key and `overwrite` is true, TryBindKey replaces this binding with
+// the new one and writes the new binding to `bindings.json`.
+// Returns true if the new binding has been applied, and a possible error
+// if failed to write to `bindings.json`.
 func TryBindKey(k, v string, overwrite bool) (bool, error) {
 	var e error
 	var parsed map[string]interface{}
@@ -303,8 +306,12 @@ func TryBindKey(k, v string, overwrite bool) (bool, error) {
 
 		BindKey(k, v, Binder["buffer"])
 
-		txt, _ := json.MarshalIndent(parsed, "", "    ")
-		return true, ioutil.WriteFile(filename, append(txt, '\n'), 0644)
+		if overwrite {
+			txt, _ := json.MarshalIndent(parsed, "", "    ")
+			return true, ioutil.WriteFile(filename, append(txt, '\n'), 0644)
+		}
+
+		return true, nil
 	}
 	return false, e
 }
